requirements.txt
python-dotenv==0.20.0
requests==2.28.1
asyncio==3.4.3
aiohttp==3.9.4
websocket-client==1.7.0
numpy==1.26.4
pandas==2.2.2 
Backtesting==0.3.3
aiogram==3.4.1
python-dotenv==1.0.0

#websocket-client==1.6.1
#numpy==1.21.6 
#pandas==1.1.5
#Backtesting==0.3.3